requirements.txt
cassandra-driver==3.7.1
six==1.10.0
